Added show reply count state to thread Moved renderMessage to <MessageItem/> Added missing 'join/leave' group function to mobile ui
