+++ components/MessageBoard/MessageList.tsx
@@ -1,16 +1,12 @@
 import { GraphQLResult } from "@aws-amplify/api"
-import { RouteProp, useNavigation } from "@react-navigation/native"
+import { RouteProp } from "@react-navigation/native"
 import { StackNavigationProp } from "@react-navigation/stack"
 import { API } from "aws-amplify"
 import GRAPHQL_AUTH_MODE from "aws-amplify-react-native"
-import { convertFromRaw } from "draft-js"
-import { stateToHTML } from "draft-js-export-html"
-import moment from "moment"
 import React, { useEffect, useState } from "react"
 import { ActivityIndicator, FlatList, Text, TouchableOpacity, View } from "react-native"
 import Observable from "zen-observable-ts"
 import { JCState } from "../../components/JCComponent/JCComponent"
-import ProfileImage from "../../components/ProfileImage/ProfileImage"
 import {
   GetMessageQuery,
   MessagesByRoomQuery,
@@ -24,11 +20,11 @@ import {
   onCreateReply,
 } from "../../src/graphql-custom/messages"
 import MessageInput from "./MessageInput"
-import MessageUtils from "./MessageUtils"
+import MessageItem from "./MessageItem"
 
-type Messages = NonNullable<MessagesByRoomQuery["messagesByRoom"]>["items"]
-type Message = NonNullable<Messages>[0]
-type Reply = NonNullable<NonNullable<NonNullable<Message>["replies"]>["items"]>[0]
+export type Messages = NonNullable<MessagesByRoomQuery["messagesByRoom"]>["items"]
+export type Message = NonNullable<Messages>[0]
+export type Reply = NonNullable<NonNullable<NonNullable<Message>["replies"]>["items"]>[0]
 
 interface Props {
   groupId?: string
@@ -86,8 +82,8 @@ export default function MessageList(props: Props): JSX.Element {
     }
     setState((prev) => ({ ...prev, fetchingData: false }))
   }
-  const renderMessageWithReplies = (item: Message, index: number) => {
-    //
+  const RenderMessageWithReplies = ({ item, index }: { item: Message; index: number }) => {
+    const [showCount, setShowCount] = useState(2)
     return (
       <View
         style={{
@@ -99,14 +95,14 @@ export default function MessageList(props: Props): JSX.Element {
         }}
         key={index}
       >
-        {renderMessage(item, index, false)}
-        {item?.replies?.items.slice(0, 2).map((reply, index) => {
-          return renderMessage(reply, index, true)
+        <MessageItem item={item} index={index} isReply={false} />
+        {item?.replies?.items.slice(0, showCount).map((reply, index) => {
+          return <MessageItem item={reply} index={index} isReply={true} />
         })}
         <TouchableOpacity
-          onPress={() => null} // increase by 2, need to add state for num replies
+          onPress={() => setShowCount((prev) => prev + 2)} // increase by 2, need to add state for num replies
         >
-          {item?.replies?.items?.length && item?.replies?.items?.length >= 3 ? (
+          {item?.replies?.items?.length && showCount < item?.replies?.items?.length ? (
             <Text
               style={{
                 marginLeft: 142,
@@ -116,8 +112,7 @@ export default function MessageList(props: Props): JSX.Element {
                 lineHeight: 24,
               }}
             >
-              View {item.replies.items.length - 2} more comments... (out of{" "}
-              {item?.replies?.items?.length})
+              View {item.replies.items.length - showCount} more comments...
             </Text>
           ) : null}
         </TouchableOpacity>
@@ -136,133 +131,11 @@ export default function MessageList(props: Props): JSX.Element {
       </View>
     )
   }
-  const renderMessage = (item: Message | Reply, index: number, isReply: boolean) => {
-    const showProfile = (id: string | undefined) => {
-      if (id) navigation?.push("ProfileScreen", { id: id, create: false })
-    }
-    const { replies } = props
-    const now = moment()
-    const datePosted = moment(parseInt(item?.when))
-    const daysSince = now.diff(datePosted.format("L"), "days")
-    const dateString = `${
-      daysSince === 0
-        ? "Today"
-        : daysSince === 1
-        ? "Yesterday"
-        : daysSince > 1
-        ? daysSince + " days ago"
-        : null
-    }`
-    return (
-      <View
-        key={index}
-        style={{
-          paddingHorizontal: 16,
-          paddingVertical: 32,
-          paddingBottom: !isReply ? 32 : 0,
-          marginLeft: isReply ? 64 : 0,
-          borderBottomWidth: !isReply ? 1 : 0,
-          borderBottomColor: "#E4E1E1",
-        }}
-      >
-        <View>
-          <View style={{ flexDirection: "row" }}>
-            <View style={{ flexDirection: "column" }}>
-              <TouchableOpacity
-                key={item?.id}
-                onPress={() => {
-                  showProfile(item?.author?.id)
-                }}
-              >
-                {item && "owner" in item && (
-                  <View style={{ marginRight: 16 }}>
-                    <ProfileImage size="editorLarge" user={item?.owner ?? null} />
-                  </View>
-                )}
-                {isReply && (
-                  <View style={{ marginLeft: -16 }}>
-                    <ProfileImage size="small2" user={item?.author?.id ?? null} />
-                  </View>
-                )}
-              </TouchableOpacity>
-            </View>
 
-            <View style={{ flex: 1 }}>
-              <View style={{ flexDirection: "row", flex: 1 }}>
-                <View style={{ flexDirection: "column", flex: 1 }}>
-                  <Text
-                    style={{
-                      color: "#1A0706",
-                      fontFamily: "Graphik-Semibold-App",
-                      fontSize: 16,
-                      lineHeight: 24,
-                    }}
-                  >
-                    {item?.author?.given_name ?? ""} {item?.author?.family_name ?? ""}
-                  </Text>
-                  <Text
-                    style={{
-                      color: "#483938",
-                      fontFamily: "Graphik-Regular-App",
-                      fontSize: 16,
-                      lineHeight: 24,
-                    }}
-                  >
-                    {item?.author?.currentRole ?? ""}
-                  </Text>
-                </View>
-                {item?.when && (
-                  <Text
-                    style={{
-                      color: "#6A5E5D",
-                      fontFamily: "Graphik-Regular-App",
-                      fontSize: 16,
-                      lineHeight: 24,
-                    }}
-                  >
-                    {dateString}
-                  </Text>
-                )}
-              </View>
-
-              <div
-                style={{
-                  fontFamily: "Graphik-Regular-App",
-                  fontWeight: "normal",
-                  fontSize: 15,
-                  lineHeight: "24px",
-                  color: "#1A0706",
-                  marginBlockEnd: 0,
-                }}
-                dangerouslySetInnerHTML={{
-                  __html: convertCommentFromJSONToHTML(item?.content ?? null),
-                }}
-              />
-
-              {item?.attachment ? <View>{MessageUtils.renderFileDownloadBadge(item)}</View> : null}
-            </View>
-          </View>
-        </View>
-      </View>
-    )
-  }
-  const convertCommentFromJSONToHTML = (text: string | null) => {
-    const errorMarkdown = "<div>" + "Message" + " Can't Be Displayed</div>"
-
-    if (!text) return errorMarkdown
-
-    try {
-      return stateToHTML(convertFromRaw(JSON.parse(text)))
-    } catch (e) {
-      console.error(e)
-      return errorMarkdown
-    }
-  }
   const messagesLoader = () => {
     if (state.nextToken) return <ActivityIndicator animating color="#F0493E" />
     return null
   }
-  const navigation = useNavigation<StackNavigationProp<any, any>>()
   const flatListRef = React.createRef<FlatList<any>>()
   useEffect(() => {
     const messageSub = (
@@ -388,9 +261,8 @@ export default function MessageList(props: Props): JSX.Element {
       <FlatList
         ref={flatListRef}
         scrollEnabled
-        renderItem={({ item, index }) => renderMessageWithReplies(item, index)}
+        renderItem={({ item, index }) => <RenderMessageWithReplies item={item} index={index} />}
         data={state.messages}
-        onEndReached={!state.fetchingData ? () => getMoreMessages() : undefined}
         refreshing={state.fetchingData}
         onEndReachedThreshold={0.1}
       />
